feat: renamed sequencer submodule to bettername 'service', splitting up previous notion of main to now be 'treeticker'
