+++ beams/bin/main.py
@@ -16,7 +16,7 @@ DESCRIPTION = __doc__
 COMMAND_TO_MODULE = {
     "run": "run",
     "validate": "validate",
-    "gen_test_ioc": "gen_test_ioc",
+    "gen_test_ioc": "gen_test_ioc"
 }
