+++ beams/service/helpers/enum.py
