+++ beams/service/helpers/timer.py
