+++ beams/service/remote_calls/heartbeat.proto
