+++ beams/service/tree_ticker.py
@@ -1,0 +1,96 @@
+"""
+Logic for `superscore run` main
+"""
+
+import logging
+import time
+from functools import partial
+from pathlib import Path
+
+from py_trees.console import read_single_keypress
+from py_trees.display import unicode_blackboard, unicode_tree
+from py_trees.trees import BehaviourTree
+from py_trees.visitors import SnapshotVisitor
+
+from beams.logging import LoggingVisitor
+from beams.tree_config import get_tree_from_path
+
+logger = logging.getLogger(__name__)
+
+
+def snapshot_post_tick_handler(
+    snapshot_visitor: SnapshotVisitor,
+    show_tree: bool,
+    show_blackboard: bool,
+    behaviour_tree: BehaviourTree,
+) -> None:
+    """
+    Print data about the part of the tree visited.
+    Does not log, to allow use of color codes.
+    ``snapshot_visitor`` keeps track of visited nodes and their status
+
+    Args:
+        snapshot_handler: gather data about the part of the tree visited
+        behaviour_tree: tree to gather data from
+    """
+    if show_tree:
+        print(
+            "\n"
+            + unicode_tree(
+                root=behaviour_tree.root,
+                visited=snapshot_visitor.visited,
+                previously_visited=snapshot_visitor.previously_visited,
+                show_status=True
+            )
+        )
+
+    if show_blackboard:
+        print(unicode_blackboard())
+
+
+def tick_tree(tree: BehaviourTree, interactive: bool, tick_delay: float):
+    tree.tick()
+    if interactive:
+        read_single_keypress()
+    else:
+        time.sleep(tick_delay)
+
+
+def tree_tick_work_func(
+    filepath: str,
+    tick_count: int,
+    tick_delay: float,
+    interactive: bool,
+    show_node_status: bool,
+    show_tree: bool,
+    show_blackboard: bool,
+):
+    logger.info(f"Running behavior tree at {filepath}")
+    # grab config
+    fp = Path(filepath).resolve()
+    if not fp.is_file():
+        raise ValueError("Provided filepath is not a file")
+
+    tree = get_tree_from_path(fp)
+    tree.visitors.append(LoggingVisitor(print_status=show_node_status))
+
+    snapshot_visitor = SnapshotVisitor()
+    tree.add_post_tick_handler(
+        partial(snapshot_post_tick_handler,
+                snapshot_visitor,
+                show_tree,
+                show_blackboard)
+    )
+    tree.setup()
+    if tick_count <= 0:
+        while True:
+            try:
+                tick_tree(tree, interactive, tick_delay)
+            except KeyboardInterrupt:
+                break
+    else:
+        for _ in range(tick_count):
+            try:
+                tick_tree(tree, interactive, tick_delay)
+            except KeyboardInterrupt:
+                break
