+++ beams/tests/test_sequencer_state.py
@@ -1,8 +1,8 @@
-from beams.sequencer.remote_calls.sequencer_pb2 import (MessageType,
+from beams.service.remote_calls.sequencer_pb2 import (MessageType,
                                                         RunStateType,
                                                         SequenceType,
                                                         TickStatus)
-from beams.sequencer.state import SequencerState, SequencerStateVariable
+from beams.service.state import SequencerState, SequencerStateVariable
 
 
 class TestTask:
